added debug output
